Updated horizontal filter list styling
Updated horizontal filter list styling to remove the in-text style that
was used to create a space at the end of the list, preventing the last
list item from being partially faded. This is now applied to
filter-options in styles.css. Also added placeholder links for the two
icons next to the filter list.